feat: generqalize
