+++ frb_codegen/src/library/codegen/parser/mir/parser/function/mod.rs
@@ -1,16 +1,14 @@
 use crate::codegen::ir::hir::flat::function::HirFlatFunction;
 use crate::codegen::ir::hir::flat::struct_or_enum::HirFlatStruct;
 use crate::codegen::ir::mir::func::MirFunc;
-use crate::codegen::ir::misc::skip::IrSkip;
+use crate::codegen::ir::misc::skip::{IrSkip, IrValueOrSkip};
 use crate::codegen::parser::mir::internal_config::ParserMirInternalConfig;
-use crate::codegen::parser::mir::parser::function::func_or_skip::MirFuncOrSkip;
 use crate::codegen::parser::mir::parser::ty::TypeParser;
 use crate::codegen::parser::mir::ParseMode;
 use itertools::{concat, Itertools};
 use std::collections::HashMap;
 
 pub(crate) mod auto_accessor;
-pub(crate) mod func_or_skip;
 pub(crate) mod real;
 
 pub(crate) fn parse(
@@ -24,7 +22,7 @@ pub(crate) fn parse(
         real::parse(src_fns, type_parser, config, parse_mode)?,
         auto_accessor::parse(config, src_structs, type_parser, parse_mode)?,
     ]);
-    let (funcs, skips) = MirFuncOrSkip::split(items);
+    let (funcs, skips) = IrValueOrSkip::split(items);
     let funcs = sort_and_add_func_id(funcs);
     Ok((funcs, skips))
 }
